+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.0] - 2026-01-06
+## [0.8.2] - 2026-01-08
 
 ### Added
 - Full backup system for encrypted content (Closes #9)
@@ -221,7 +221,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Responsive masonry grid layout
 - Lightbox photo viewer
 
-[0.9.0]: https://github.com/Mistress-Lukutar/Synth-gallery/compare/v0.8.1...v0.9.0
+[0.8.2]: https://github.com/Mistress-Lukutar/Synth-gallery/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/Mistress-Lukutar/Synth-gallery/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/Mistress-Lukutar/Synth-gallery/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/Mistress-Lukutar/Synth-gallery/compare/v0.6.2...v0.7.0